tmp commit

+++ src/main.rs
@@ -27,7 +27,7 @@ fn main() {
         .run();
 }
 
-pub const CHUNK_SIZE: usize = 128;
+pub const CHUNK_SIZE: u32 = 128;
 pub const CHUNK_SCALE: usize = 3;
 pub const BYTES_PER_PIXEL: usize = 4;
